+++ backend/canisters/multi_user/impl/src/updates/create_community.rs
@@ -7,8 +7,8 @@ use oc_error_codes::OCErrorCode;
 use std::collections::HashSet;
 use types::{CanisterId, CommunityId, OCResult};
 use user_canister::create_community::{Response::*, *};
-use user_state::COMMUNITY_CREATION_LIMIT;
-use user_state::User;
+use user_core::COMMUNITY_CREATION_LIMIT;
+use user_core::User;
 use utils::document::{validate_avatar, validate_banner};
 use utils::text_validation::{
     NameValidationError, RulesValidationError, validate_channel_name, validate_community_name, validate_description,