+++ backend/canisters/user/impl/src/queries/updates.rs
@@ -11,11 +11,9 @@ fn updates(args: Args) -> Response {
 
 fn updates_impl(updates_since: TimestampMillis, state: &RuntimeState) -> Response {
     // `now` is only read if there are updates, so that caching works effectively
-    match state
-        .data
-        .user
-        .updates(updates_since, state.env.canister_id().into(), || state.env.now())
-    {
+    match user_core::queries::updates(&state.data.user, updates_since, state.env.canister_id().into(), || {
+        state.env.now()
+    }) {
         Some(result) => Success(result),
         None => SuccessNoUpdates,
     }